Display email metrics and loading progress bar
