+++ app/controllers/reports/line-items/email.js
@@ -1,0 +1,5 @@
+import Controller from '@ember/controller';
+
+export default Controller.extend({
+
+});
