+++ app/routes/reports/line-items/email/metrics.js
@@ -1,7 +1,7 @@
 import Route from '@ember/routing/route';
 import { RouteQueryManager } from 'ember-apollo-client';
 
-import query from 'leads-manage/gql/queries/lead-report/email-metrics';
+import query from 'leads-manage/gql/queries/reports/line-items/email/metrics';
 
 export default Route.extend(RouteQueryManager, {
   queryParams: {
@@ -19,13 +19,8 @@ export default Route.extend(RouteQueryManager, {
    */
   model({ sortBy, ascending }) {
     const hash = this.modelFor('reports.line-items').get('hash');
-
     const sort = { field: sortBy, order: ascending ? 1 : -1 };
     const variables = { hash, sort };
-
-    console.info(variables);
-    return {};
-
-    // return this.get('apollo').watchQuery({ query, variables, fetchPolicy: 'network-only' }, 'reportEmailMetrics');
+    return this.get('apollo').watchQuery({ query, variables, fetchPolicy: 'network-only' }, 'emailLineItemMetricsReport');
   },
 });
